requirements.txt
python-telegram-bot==20.7
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
yt-dlp==2023.12.30
validators==0.22.0
Pillow==10.1.0
Flask==3.0.0
gunicorn==21.2.0